requirements.txt
discord.py
pymongo
gitpython